[Feat] 홈-기록없음 -> 온보딩뷰로 전환 #5
